sign-to-text-ai/backend/requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# CORS and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Text-to-Speech (TTS)
gTTS>=2.5.3

# Speech-to-Text (STT) - Whisper
faster-whisper>=1.0.0  # Recommended: faster and optimized
# openai-whisper>=20231117  # Alternative: uncomment if faster-whisper fails

# Audio Processing
pydub>=0.25.1  # Required for audio format conversion (M4A, MP3, etc.)
librosa>=0.10.0  # Alternative audio loader (better error handling)
soundfile>=0.12.0  # Required by librosa for audio I/O
ffmpeg-python>=0.2.0  # Optional: for advanced audio processing

# Image Processing
Pillow>=11.0.0

# Database (Python 3.12 compatible - all have pre-built wheels!)
sqlalchemy>=2.0.35
asyncpg>=0.29.0
alembic>=1.14.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
aiofiles>=23.2.1
redis>=5.0.1

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.12.1
flake8>=6.1.0

# Additional dependencies (installed during setup)
numpy>=1.26.4
aiosqlite>=0.19.0
email-validator>=2.0.0
dnspython>=2.0.0

# Note: ML libraries will be added in Phase 2 (requirements-ml.txt)
# This keeps the base requirements light for development
